requirements.txt
llama-cloud-services>=0.1.0
python-dotenv>=0.19.0
pydantic>=2.0.0 